feat: spawngd packaging
