+++ src/spawngd/__init__.py
@@ -1,0 +1,5 @@
+from .spawngd import SpawnGD
+from .spawngd_ms import SpawnGD_MS
+
+__version__ = "0.1.0"
+__all__ = ["SpawnGD", "spawngd", "SpawnGD_MS", "spawngd_ms"]
